Created clients receive messages via handle_message
