+++ menuflow/matrix.py
@@ -1,6 +1,98 @@
-from mautrix.client import Client as MatrixClient, SyncStream
+from mautrix.client import Client as MatrixClient
+from mautrix.types import Membership, MessageEvent, StrippedStateEvent
+
+from .flow import Flow
+from .user import User
 
 
 class MenuFlowMatrixClient(MatrixClient):
     def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
+
+    async def handle_invite(self, evt: StrippedStateEvent) -> None:
+        # if evt.state_key == self.id and evt.content.membership == Membership.INVITE:
+        #     await self.client.join_room(evt.room_id)
+        pass
+
+    async def handle_message(self, evt: StrippedStateEvent) -> None:
+        self.log.debug(f"incoming message {evt.content}")
+        # Ignore bot messages
+
+        if evt.sender in self.config["menuflow.users_ignore"] or evt.sender == self.client.mxid:
+            return
+
+        try:
+            user = await User.get_by_user_id(user_id=evt.sender)
+            user.flow = self.menuflow
+        except Exception as e:
+            self.log.exception(e)
+            return
+
+        if not user:
+            return
+
+        await self.algorithm(user=user, evt=evt)
+
+    async def algorithm(self, user: User, evt: MessageEvent) -> None:
+        """If the user is in the input state, then set the variable to the user's input,
+        and if the node has an output connection, then update the menu to the output connection.
+        Otherwise, run the node and update the menu to the output connection.
+        If the node is an input node and the user is not in the input state,
+        then show the message and update the menu to the node's id and set the state to input.
+        If the node is a message node, then show the message and if the node has an output connection,
+        then update the menu to the output connection and run the algorithm again
+
+        Parameters
+        ----------
+        user : User
+            User - the user object
+        evt : MessageEvent
+            The event that triggered the algorithm.
+
+        Returns
+        -------
+            The return value is the result of the last expression in the function body.
+
+        """
+
+        # This is the case where the user is in the input state.
+        # In this case, the variable is set to the user's input, and if the node has an output connection,
+        # then the menu is updated to the output connection.
+        # Otherwise, the node is run and the menu is updated to the output connection.
+        if user.state == "input":
+            await user.set_variable(user.node.variable, evt.content.body)
+
+            if user.node.o_connection:
+                await user.update_menu(context=user.node.o_connection)
+            else:
+                o_connection = await user.node.run(user=user)
+                await user.update_menu(context=o_connection)
+
+        # This is the case where the user is not in the input state and the node is an input node.
+        # In this case, the message is shown and the menu is updated to the node's id and the state is set to input.
+        if user.node.type == "input" and user.state != "input":
+            await user.node.show_message(
+                variables=user._variables, room_id=evt.room_id, client=evt.client
+            )
+            self.log.debug(f"Input {user.node}")
+            await user.update_menu(context=user.node.id, state="input")
+            return
+
+        # Showing the message and updating the menu to the output connection.
+        if user.node.type == "message":
+            await user.node.show_message(
+                variables=user._variables, room_id=evt.room_id, client=evt.client
+            )
+            self.log.debug(f"Message {user.node}")
+
+            if user.node.o_connection is None:
+                return
+
+            await user.update_menu(context=user.node.o_connection)
+
+        if user.node.type == "http_request":
+            self.log.debug(f"HTTPRequest {user.node}")
+
+            await user.node.request(user=user, session=evt.client.api.session)
+
+        await self.algorithm(user=user, evt=evt)
