+++ menuflow/utils/base_logger.py
@@ -1,0 +1,7 @@
+from logging import getLogger
+
+from mautrix.util.logging import TraceLogger
+
+
+class BaseLogger:
+    log: TraceLogger = getLogger("menuflow.node")
